requirements.txt
allure-behave==2.12.0
allure-python-commons==2.12.0
async-generator==1.10
attrs==22.2.0
behave==1.2.6
certifi==2022.12.7
colorlog==6.7.0
charset-normalizer==3.0.1
exceptiongroup==1.1.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
outcome==1.2.0
packaging==23.0
parse==1.19.0
parse-type==0.6.0
pluggy==1.0.0
pydantic==1.10.5
PySocks==1.7.1
pytest==7.2.1
python-dotenv==0.21.1
requests==2.28.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.5.0
urllib3==1.26.14
webdriver-manager==3.8.5
wsproto==1.2.0
pytest-playwright
allure-pytest
jsonschema~=4.20.0
faker~=21.0.0
hvac~=2.0.0
pytest-xdist
playwright==1.43.0
ruff==0.2.1
pre-commit==3.6.1
pytest-check==2.2.4
xmldiff==2.6.3
lxml==5.1.0
jsondiff==2.0.0
pytest-ordering==0.6
